requirements.txt
celery
coverage
django
djangorestframework==3.11.1
django-celery-beat==2.2.0
drf-yasg==1.20.0
gunicorn
ffmpeg
opencv-python
psycopg2-binary==2.8.5
python-decouple==3.3
pylint
pytest
Pillow
redis
whitenoise==5.1.0